- Added blur+glow shader test and scene graph test
