+++ src/renderer/scene_graph/node.hpp
