+++ src/renderer/scene_graph/spatial.hpp
@@ -8,7 +8,7 @@
 #include <tools/log.hpp>
 #include <core/enumfield.hpp>
 #include <maths/transform.hpp>
-#include <scene_graph/bound.hpp>
+#include <renderer/scene_graph/bound.hpp>
 
 namespace zap { namespace scene_graph {
     template <typename SpatialT, typename PtrT=std::unique_ptr<SpatialT>> class node;
