+++ src/renderer/scene_graph/visual.hpp
